+++ source/mdcii/mdcii/include/mdcii/gam/island.hpp
@@ -22,35 +22,9 @@
 #include <string>
 #include <vector>
 
+#include "islandhouse.hpp"
 #include "productionlist.hpp"
 #include "warehouse.hpp"
-
-struct IslandHouseData
-{
-  uint16_t id;                 // tile gaphic ID, see haeuser.cod for referene
-  uint8_t posx;                // position on island
-  uint8_t posy;                // position on island
-  uint32_t orientation : 2;    // orientation
-  uint32_t animationCount : 4; // animation step for tile
-  uint32_t islandNumber : 8;   // the island the field is part of
-  uint32_t cityNumber : 3;     // the city the field is part of
-  uint32_t randomNumber : 5;   // random number, what for?
-  uint32_t playerNumber : 4;   // the player that occupies this field
-};
-
-class IslandHouse
-{
-public:
-  IslandHouse()
-  {
-  }
-  explicit IslandHouse(uint8_t* data, uint32_t length, const char* name);
-
-private:
-  std::string name;
-  std::vector<IslandHouseData> islandHouse;
-};
-
 
 struct OreMountainData // Erzberg
 {
